Removed BT
Looks like we have to make due without BT. We're running out of heap space.
I tried a bunch to save heap space. The biggest thief is the jpg buffer. We get the best savings by removing it and sending the buffer from the framebuffer directly. But then we're seeing some performance issues instead.